+++ examples/main.cayenne.py
@@ -34,6 +34,7 @@ class ExampleDatalogger(TerkinDatalogger):
 def main():
     """Start the data logger application."""
     datalogger = ExampleDatalogger(settings)
+    datalogger.setup()
     datalogger.start()